feat: database migrations 051-056 & schema hardening
- Migration 051: fix model catalog Ollama names
- Migration 052: document unique content hash constraint
- Migration 053: update thinking models configuration
- Migration 054: FK constraints, kg_entity_documents index, llm_jobs index
- Migration 055: telegram bot capabilities schema
- Migration 056: schema cleanup - duplicate indexes, FK indexes, telegram session cleanup
- PostgreSQL tuning: 1GB shared_buffers, 3GB effective_cache, zstd WAL compression
- Autovacuum tuning for Jetson workloads
- Updated existing migrations with missing indexes and constraints

+++ services/postgres/init/009_documents_schema.sql
@@ -214,8 +214,8 @@ BEGIN
         COUNT(*) FILTER (WHERE status = 'indexed'),
         COUNT(*) FILTER (WHERE status = 'pending'),
         COUNT(*) FILTER (WHERE status = 'failed'),
-        COALESCE(SUM(chunk_count) FILTER (WHERE status = 'indexed'), 0),
-        COALESCE(SUM(file_size) FILTER (WHERE status != 'deleted'), 0),
+        COALESCE(SUM(chunk_count) FILTER (WHERE status = 'indexed'), 0)::bigint,
+        COALESCE(SUM(file_size) FILTER (WHERE status != 'deleted'), 0)::bigint,
         (
             SELECT jsonb_object_agg(
                 COALESCE(dc.name, 'Unkategorisiert'),
